requirements.txt
MouseInfo==0.0.4
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
PyAutoGUI==0.9.47
PyGetWindow==0.0.7
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
